Created components for menu and about pages
