+++ frontend/src/pages/menu/index.jsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { Link } from "react-router-dom";
+
 import { MenuDataAction, addCart } from "../../redux/actions/cart-actions";
-import { FaCartPlus } from "react-icons/fa6";
+
 import Loader from "../../components/loader";
-import { MdOutlineRemoveRedEye } from "react-icons/md";
-import { paths } from "../../utils/paths";
+
+import AppsForOrder from "../../components/apps-for-order";
+import ItemCard from "../../components/ItemCard";
 
 const Menu = () => {
   const dispatch = useDispatch();
@@ -54,16 +55,7 @@ const Menu = () => {
   const handleadd = (item) => {
     dispatch(addCart(item));
   };
-  const imagesData = [
-    "/Images/deliveroo.png",
-    "/Images/justEat.png",
-    "/Images/foodpanda.png",
-    "/Images/didiFood.png",
-    "/Images/instacart.png",
-    "/Images/doordash.png",
-    "/Images/uberEats.png",
-    "/Images/grubhub.png",
-  ];
+
   return loading ? (
     <Loader />
   ) : (
@@ -110,53 +102,18 @@ const Menu = () => {
 
         <div className="grid px-4 gap-12 md:gap-6 grid-cols-1 md:grid-cols-4">
           {Items?.map((item, i) => (
-            <div key={i} className="rounded-t-xl overflow-hidden">
-              <img src={item.image} alt="menu" className="w-full h-[230px]" />
-              <div className=" px-[30px] py-[30px] border border-t-0 border-[#DBDFD0] rounded-b-xl">
-                <p className=" text-lg font-medium text-third text-center leading-7 ">
-                  ${item.price}
-                </p>
-                <p className="my-[22px]  text-lg font-medium text-third text-center">
-                  {item.name}
-                </p>
-                <div className="flex gap-3 justify-center text-xl text-primary font-semibold">
-                  <Link to={`${paths.itemDetail}${item._id}`}>
-                    <MdOutlineRemoveRedEye />
-                  </Link>
-                  <button onClick={() => handleadd(item)}>
-                    <FaCartPlus />
-                  </button>
-                </div>
-              </div>
-            </div>
+            <ItemCard
+              image={item.image}
+              price={item.price}
+              name={item.name}
+              id={item._id}
+              item={item}
+              handleadd={handleadd}
+            />
           ))}
         </div>
       </div>
-      <div className="bg-[#e9e9e947]">
-        <div className="main-container py-[90px] flex flex-col md:flex-row gap-10">
-          <div className="w-full md:w-[40%] flex flex-col justify-center">
-            <h2 className="text-5xl font-bold text-secondary leading-tight">
-              You can order through apps
-            </h2>
-            <p className="text-base text-third leading-7 mt-[30px]">
-              Lorem ipsum dolor sit amet consectetur adipiscing elit enim
-              bibendum sed et aliquet aliquet risus tempor semper.
-            </p>
-          </div>
-          <div className="w-full md:w-[60%] flex flex-wrap gap-10 justify-center">
-            {imagesData?.map((value, i) => (
-              <div
-                key={i}
-                className={`bg-white h-[90px] flex justify-center items-center rounded-xl ${
-                  i + 1 === (4 || 5 || 6) ? "w-[280px]" : "w-[230px]"
-                }`}
-              >
-                <img src={value} alt="menu" />
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
+      <AppsForOrder />
     </>
   );
 };
